requirements.txt
Django==4.0
psycopg2-binary==2.9.1
Pillow==9.3.0
djangorestframework==3.14.0